octofit-tracker/backend/requirements.txt
Django==4.1.7
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
django-allauth==0.51.0
django-cors-headers==4.5.0
dj-rest-auth==2.2.6
django-filter==23.3
Pillow==9.5.0
python-dotenv==1.0.0
psycopg2-binary==2.9.9
whitenoise==6.5.0
drf-yasg==1.21.7